dir reorg - move menu panels into `BookNav`, new 'feaures (controllers?) dir
