+++ src/BookNavigator/features/bookmarks/ia-bookmarks.js
@@ -1,7 +1,7 @@
 /* eslint-disable compat/compat */
 import { render, nothing } from 'lit-html';
 import { LitElement, html, css } from 'lit-element';
-import buttonStyles from './styles/button-base.js';
+import buttonStyles from '../../styles/button-base.js';
 
 const api = {
   endpoint: '/services/bookmarks.php',
