+++ src/BookNavigator/providers/visual-adjustments.js
@@ -1,7 +1,7 @@
 import { html } from 'lit-element';
 
 /* define subpanel */
-import { IABookVisualAdjustments } from '../../iaux-book-visual-adjustments/ia-book-visual-adjustments';
+import { IABookVisualAdjustments } from '../menu-panels/visual-adjustments/ia-book-visual-adjustments';
 customElements.define('ia-book-visual-adjustments', IABookVisualAdjustments);
 
 const visualAdjustmentOptions = [{
